pyproject.toml
[project]
name = "pdftools"
version = "0.2.0"
description = "Cross-platform PDF utility with native-feeling PySide6 UI and optional SaaS backend"
authors = [{ name = "pdfTools Team" }]
requires-python = ">=3.9"
dependencies = [
    "pypdf>=5.0",
    "PySide6>=6.5,<6.6",
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "Pillow>=10.0",
]
readme = "AGENTS.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "black>=24.4",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
pdftools-gui = "pdftools.gui.main:main"
pdftools-server = "pdftools.server.app:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
target-version = ["py39"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra"
